fix: 청소코드 인증 경로 CleaningCodeController -> CheckController 이동
