+++ src/main/java/com/dormclean/dorm_cleaning_management/service/CheckService.java
@@ -1,6 +1,7 @@
 package com.dormclean.dorm_cleaning_management.service;
 
 import com.dormclean.dorm_cleaning_management.dto.CheckRequestDto;
+import com.dormclean.dorm_cleaning_management.dto.CleaningCodeDto;
 
 public interface CheckService {
     void checkIn(CheckRequestDto dto);
@@ -8,4 +9,6 @@ public interface CheckService {
     void checkOut(CheckRequestDto dto);
 
     void cleanCheck(CheckRequestDto dto);
+
+    void useCleaningCode(CleaningCodeDto dto);
 }
